week-3/mnist-train/requirements.txt
typer==0.6.1
tqdm==4.64.1
wandb==0.13.3
PyYAML
numpy
pytest
pytest-cov
pandas
